fix: move resolveArgs to CLI instance
